requirement.txt
groq
streamlit
faiss-cpu 
numpy
requests
python-dotenv